refactor: split redisconnector into modules
